feat: GET /records

+++ backend/proficiv/domain/records/schema.py
@@ -8,7 +8,7 @@ from proficiv.entity import ActionID, RecordID, SubjectID
 class IRecord(BaseModel):
     record_id: RecordID
     subject_id: SubjectID
-    fps: int = Field(description="Frames per second; 一秒間あたりのサンプリングレート")
+    fps: float = Field(description="Frames per second; 一秒間あたりのサンプリングレート")
     head_camera_video_url: HttpUrl
